add ddl model
