+++ inlong-sort/sort-common/src/main/java/org/apache/inlong/sort/ddl/DdlConstants.java
@@ -1,0 +1,5 @@
+package org.apache.inlong.sort.ddl;
+
+public class DdlConstants {
+
+}
