+++ inlong-sort/sort-common/src/main/java/org/apache/inlong/sort/ddl/Utils/OperationUtils.java
@@ -1,0 +1,7 @@
+package org.apache.inlong.sort.ddl.Utils;
+
+
+public class OperationUtils {
+
+
+}
