+++ inlong-sort/sort-formats/format-json/src/main/java/org/apache/inlong/sort/formats/json/canal/CanalJson.java
@@ -29,7 +29,7 @@ import org.apache.inlong.sort.ddl.operations.Operation;
 
 @Builder
 @Data
-@JsonTypeName("CanalJson")
+@JsonTypeName("canalJson")
 public class CanalJson {
 
     @JsonProperty("data")
